+++ packages/bloc/test/helpers/complex/complex_bloc.dart
@@ -13,7 +13,7 @@ class ComplexBloc extends Bloc<ComplexEvent, ComplexState> {
     Stream<ComplexEvent> events,
     Function(ComplexEvent) next,
   ) {
-    return (events as Observable<ComplexEvent>).switchMap(next);
+    return events.switchMap(next);
   }
 
   @override
@@ -33,20 +33,6 @@ class ComplexBloc extends Bloc<ComplexEvent, ComplexState> {
 
   @override
   Stream<ComplexState> transformStates(Stream<ComplexState> states) {
-    return (states as Observable<ComplexState>)
-        .debounceTime(Duration(milliseconds: 50));
+    return states.debounceTime(Duration(milliseconds: 50));
   }
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is ComplexBloc &&
-          runtimeType == other.runtimeType &&
-          initialState == other.initialState;
-
-  @override
-  int get hashCode =>
-      initialState.hashCode ^
-      mapEventToState.hashCode ^
-      transformEvents.hashCode;
 }