+++ packages/bloc_test/lib/src/bloc_test.dart
@@ -1,9 +1,8 @@
 import 'dart:async';
 
 import 'package:bloc/bloc.dart';
-import 'package:bloc_test/bloc_test.dart';
 import 'package:meta/meta.dart';
-import 'package:test/test.dart';
+import 'package:test/test.dart' as test;
 
 /// Creates a new [bloc]-specific test case with the given [description].
 /// [blocTest] will handle asserting that the [bloc] emits the [expect]ed
@@ -14,11 +13,17 @@ import 'package:test/test.dart';
 /// [build] should be used for all [bloc] initialization and preparation
 /// and must return the [bloc] under test.
 ///
-/// [act] is an optional callback which will be invoked with the [bloc] under test
-/// and should be used to `add` events to the bloc.
+/// [act] is an optional callback which will be invoked with the [bloc] under
+/// test and should be used to `add` events to the [bloc].
 ///
-/// [expect] is an `Iterable<State>` which the [bloc]
+/// [wait] is an optional `Duration` which can be used to wait for
+/// async operations within the [bloc] under test such as `debounceTime`.
+///
+/// [expect] is an `Iterable` of matchers which the [bloc]
 /// under test is expected to emit after [act] is executed.
+///
+/// [verify] is an optional callback which is invoked after [expect]
+/// and can be used for additional non-bloc related assertions.
 ///
 /// ```dart
 /// blocTest(
@@ -39,16 +44,64 @@ import 'package:test/test.dart';
 ///   expect: [0],
 /// );
 /// ```
+///
+/// [blocTest] can also be used to [verify] internal bloc functionality.
+///
+/// ```dart
+/// blocTest(
+///   'CounterBloc emits [0, 1] when CounterEvent.increment is added',
+///   build: () => CounterBloc(),
+///   act: (bloc) => bloc.add(CounterEvent.increment),
+///   expect: [0, 1],
+///   verify: () async {
+///     verify(repository.someMethod(any)).called(1);
+///   }
+/// );
+/// ```
+///
+/// [blocTest] can also be used to wait for async operations like `debounceTime`
+/// by optionally providing a `Duration` to [wait].
+///
+/// ```dart
+/// blocTest(
+///   'CounterBloc emits [0, 1] when CounterEvent.increment is added',
+///   build: () => CounterBloc(),
+///   act: (bloc) => bloc.add(CounterEvent.increment),
+///   wait: const Duration(milliseconds: 300),
+///   expect: [0, 1],
+/// );
+/// ```
+///
+/// **Note:** when using [blocTest] with state classes which don't override
+/// `==` and `hashCode` you can provide an `Iterable` of matchers instead of
+/// explicit state instances.
+///
+/// ```dart
+/// blocTest(
+///  'emits [StateA, StateB] when MyEvent is added',
+///  build: () => MyBloc(),
+///  act: (bloc) => bloc.add(MyEvent()),
+///  expect: [isA<StateA>(), isA<StateB>()],
+/// );
+/// ```
 @isTest
 void blocTest<B extends Bloc<Event, State>, Event, State>(
   String description, {
-  @required B build(),
-  @required Iterable<State> expect,
+  @required B Function() build,
+  @required Iterable expect,
   Future<void> Function(B bloc) act,
+  Duration wait,
+  Future<void> Function() verify,
 }) {
-  test(description, () async {
+  test.test(description, () async {
     final bloc = build();
+    final states = <State>[];
+    final subscription = bloc.listen(states.add);
     await act?.call(bloc);
-    await emitsExactly(bloc, expect);
+    if (wait != null) await Future.delayed(wait);
+    await bloc.close();
+    test.expect(states, expect);
+    await subscription.cancel();
+    await verify?.call();
   });
 }