+++ packages/flutter_bloc/lib/src/bloc_listener.dart
@@ -1,35 +1,42 @@
 import 'dart:async';
 
-import 'package:flutter/widgets.dart';
-
 import 'package:bloc/bloc.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:provider/provider.dart';
+import 'package:flutter/widgets.dart';
+import 'package:provider/single_child_widget.dart';
 
-/// Signature for the [listener] function which takes the `BuildContext` along with the [bloc] [state]
-/// and is responsible for executing in response to [state] changes.
+import '../flutter_bloc.dart';
+
+/// Mixin which allows `MultiBlocListener` to infer the types
+/// of multiple [BlocListener]s.
+mixin BlocListenerSingleChildWidget on SingleChildWidget {}
+
+/// Signature for the [listener] function which takes the `BuildContext` along
+/// with the [bloc] [state] and is responsible for executing in response to
+/// [state] changes.
 typedef BlocWidgetListener<S> = void Function(BuildContext context, S state);
 
-/// Signature for the [condition] function which takes the previous [state] and the current [state]
-/// and is responsible for returning a [bool] which determines whether or not to call [BlocWidgetListener]
-/// of [BlocListener] with the current [state].
+/// Signature for the [condition] function which takes the previous [state]
+/// and the current [state] and is responsible for returning a [bool] which
+/// determines whether or not to call [BlocWidgetListener] of [BlocListener]
+/// with the current [state].
 typedef BlocListenerCondition<S> = bool Function(S previous, S current);
 
 /// {@template bloclistener}
-/// Takes a [BlocWidgetListener] and an optional [bloc]
-/// and invokes the [listener] in response to [state] changes in the [bloc].
-/// It should be used for functionality that needs to occur only in response to a [state] change
-/// such as navigation, showing a [SnackBar], showing a [Dialog], etc...
-/// The [listener] is guaranteed to only be called once for each [state] change unlike the
-/// [builder] in [BlocBuilder].
+/// Takes a [BlocWidgetListener] and an optional [bloc] and invokes
+/// the [listener] in response to [state] changes in the [bloc].
+/// It should be used for functionality that needs to occur only in response to
+/// a [state] change such as navigation, showing a [SnackBar], showing
+/// a [Dialog], etc...
+/// The [listener] is guaranteed to only be called once for each [state] change
+/// unlike the [builder] in [BlocBuilder].
 ///
-/// If the [bloc] parameter is omitted, [BlocListener] will automatically perform a lookup using
-/// [BlocProvider] and the current `BuildContext`.
+/// If the [bloc] parameter is omitted, [BlocListener] will automatically
+/// perform a lookup using [BlocProvider] and the current `BuildContext`.
 ///
 /// ```dart
 /// BlocListener<BlocA, BlocAState>(
 ///   listener: (context, state) {
-///     /// do stuff here based on BlocA's state
+///     // do stuff here based on BlocA's state
 ///   },
 ///   child: Container(),
 /// )
@@ -41,7 +48,7 @@ typedef BlocListenerCondition<S> = bool Function(S previous, S current);
 /// BlocListener<BlocA, BlocAState>(
 ///   bloc: blocA,
 ///   listener: (context, state) {
-///     /// do stuff here based on BlocA's state
+///     // do stuff here based on BlocA's state
 ///   },
 ///   child: Container(),
 /// )
@@ -50,16 +57,17 @@ typedef BlocListenerCondition<S> = bool Function(S previous, S current);
 /// An optional [condition] can be implemented for more granular control
 /// over when [listener] is called.
 /// The [condition] function will be invoked on each [bloc] [state] change.
-/// The [condition] takes the previous [state] handled by [BlocListener] and
-/// the current [state] of the [bloc] and must return a [bool] which determines
-/// whether or not the [listener] function will be invoked.
+/// The [condition] takes the previous [state] and current [state] and must
+/// return a [bool] which determines whether or not the [listener] function
+/// will be invoked.
 /// The previous [state] will be initialized to the [state] of the [bloc]
 /// when the [BlocListener] is initialized.
-/// [condition] is optional and if it isn't implemented, it will default to `true`.
+/// [condition] is optional and if it isn't implemented, it will default to
+/// `true`.
 ///
 /// ```dart
 /// BlocListener<BlocA, BlocAState>(
-///   condition: (previousState, state) {
+///   condition: (previous, current) {
 ///     // return true/false to determine whether or not
 ///     // to invoke listener with state
 ///   },
@@ -71,7 +79,7 @@ typedef BlocListenerCondition<S> = bool Function(S previous, S current);
 /// ```
 /// {@endtemplate}
 class BlocListener<B extends Bloc<dynamic, S>, S> extends BlocListenerBase<B, S>
-    with SingleChildCloneableWidget {
+    with BlocListenerSingleChildWidget {
   /// The widget which will be rendered as a descendant of the [BlocListener].
   final Widget child;
 
@@ -85,26 +93,11 @@ class BlocListener<B extends Bloc<dynamic, S>, S> extends BlocListenerBase<B, S>
   })  : assert(listener != null),
         super(
           key: key,
+          child: child,
           listener: listener,
           bloc: bloc,
           condition: condition,
         );
-
-  /// Clones the current [BlocListener] with a new [child] widget.
-  /// All other values, including [key], [bloc] and [listener] are preserved.
-  @override
-  BlocListener<B, S> cloneWithChild(Widget child) {
-    return BlocListener<B, S>(
-      key: key,
-      bloc: bloc,
-      listener: listener,
-      condition: condition,
-      child: child,
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) => child;
 }
 
 /// {@template bloclistenerbase}
@@ -115,7 +108,11 @@ class BlocListener<B extends Bloc<dynamic, S>, S> extends BlocListenerBase<B, S>
 /// is defined by sub-classes.
 /// {@endtemplate}
 abstract class BlocListenerBase<B extends Bloc<dynamic, S>, S>
-    extends StatefulWidget {
+    extends SingleChildStatefulWidget {
+  /// The widget which will be rendered as a descendant of the
+  /// [BlocListenerBase].
+  final Widget child;
+
   /// The [bloc] whose [state] will be listened to.
   /// Whenever the [bloc]'s [state] changes, [listener] will be invoked.
   final B bloc;
@@ -128,10 +125,13 @@ abstract class BlocListenerBase<B extends Bloc<dynamic, S>, S>
 
   /// The [BlocListenerCondition] that the [BlocListenerBase] will invoke.
   /// The [condition] function will be invoked on each [bloc] [state] change.
-  /// The [condition] takes the previous [state] and current [state] and must return a [bool]
-  /// which determines whether or not the [listener] function will be invoked.
-  /// The previous [state] will be initialized to [state] when the [BlocListenerBase] is initialized.
-  /// [condition] is optional and if it isn't implemented, it will default to `true`.
+  /// The [condition] takes the previous [state] and current [state] and must
+  /// return a [bool] which determines whether or not the [listener] function
+  /// will be invoked.
+  /// The previous [state] will be initialized to [state] when
+  /// the [BlocListenerBase] is initialized.
+  /// [condition] is optional and if it isn't implemented, it will default to
+  /// `true`.
   final BlocListenerCondition<S> condition;
 
   /// {@macro bloclistenerbase}
@@ -139,17 +139,17 @@ abstract class BlocListenerBase<B extends Bloc<dynamic, S>, S>
     Key key,
     this.listener,
     this.bloc,
+    this.child,
     this.condition,
-  }) : super(key: key);
+  }) : super(key: key, child: child);
 
-  State<BlocListenerBase<B, S>> createState() => _BlocListenerBaseState<B, S>();
-
-  /// Returns a widget based on the `BuildContext`.
-  Widget build(BuildContext context);
+  @override
+  SingleChildState<BlocListenerBase<B, S>> createState() =>
+      _BlocListenerBaseState<B, S>();
 }
 
 class _BlocListenerBaseState<B extends Bloc<dynamic, S>, S>
-    extends State<BlocListenerBase<B, S>> {
+    extends SingleChildState<BlocListenerBase<B, S>> {
   StreamSubscription<S> _subscription;
   S _previousState;
   B _bloc;
@@ -178,7 +178,7 @@ class _BlocListenerBaseState<B extends Bloc<dynamic, S>, S>
   }
 
   @override
-  Widget build(BuildContext context) => widget.build(context);
+  Widget buildWithChild(BuildContext context, Widget child) => child;
 
   @override
   void dispose() {
@@ -188,11 +188,11 @@ class _BlocListenerBaseState<B extends Bloc<dynamic, S>, S>
 
   void _subscribe() {
     if (_bloc != null) {
-      _subscription = _bloc.skip(1).listen((S state) {
+      _subscription = _bloc.skip(1).listen((state) {
         if (widget.condition?.call(_previousState, state) ?? true) {
           widget.listener(context, state);
-          _previousState = state;
         }
+        _previousState = state;
       });
     }
   }